GEP-128 Expose some internal API of flex-layout
This exposition is required as geppetto-client is using them. As we need
to switch to a ts compatible version of geppetto-client, the compiler
complains about those methods from various objects in the flex-layout
API that were not exposed. This commit exposes the ones that were
required by geppetto-client. However, a rewrite of those classes should
be done.